back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "celery>=5.3.6",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "pydantic[email]>=2.12.5",
    "pywin32>=306; platform_system == 'Windows'",
    "redis>=5.0.0",
]
